refactor: restructure PSS device architecture and simplify sensor naming conventions
